move code files to crashsimilarity package
