+++ crashsimilarity/top_similar_traces.py
@@ -1,7 +1,7 @@
 # CLI INTERFACE THAT TAKES STACK TRACE AS INPUT AND RETURNS SIMILAR STACK TRACES
-import download_data
-import crash_similarity
 import argparse
+
+from crashsimilarity import download_data, crash_similarity
 
 parser = argparse.ArgumentParser(description='Returns the top ten similar stack traces')
 parser.add_argument('--crash_id', required=True, help='crash_id corresponding to the stack trace')
